Добавлены базовые "азы" режима редактирования

+++ qlife/fieldthread.h
@@ -28,9 +28,9 @@ public:
     void randomModel(int start_i = Config::instance()->randomStartN(), int start_j = Config::instance()->randomStartM(), int i_cnt = Config::instance()->randomCountN(), int j_cnt = Config::instance()->randomCountM());
 	void clearModel();
 
-	void setStepCount(unsigned long steps) { m_stepCount = steps; }
-	unsigned long stepCount()const { return m_stepCount; }
-	static unsigned long GetInfiniteSteps() { return 0xffffffffffffffff; }
+	void setStepCount(unsigned long long steps) { m_stepCount = steps; }
+	unsigned long long stepCount()const { return m_stepCount; }
+	static unsigned long long GetInfiniteSteps() { return 0xffffffffffffffff; }
 	void setInfiniteSteps() { m_stepCount = GetInfiniteSteps(); }
 
     inline unsigned char randomBool();
@@ -42,7 +42,7 @@ protected:
     virtual void run()override;
     QSharedPointer<DynModel> m_model;
     mutable bool m_isRunning;
-	mutable unsigned long m_stepCount;
+	mutable unsigned long long m_stepCount;
 	QMutex m_mutex;
 
 signals:
